+++ src/components/AppBar/AppBar.jsx
@@ -1,5 +1,5 @@
 import { AuthNav } from 'components/AuthNav/AuthNav';
-import { LoginForm } from 'components/LoginForm/LogimForm';
+
 import { Link } from 'react-router-dom';
 import logo from '../../images/logo.svg';
 import { StyledHeader } from './AppBar.styled';
@@ -11,7 +11,6 @@ export const AppBar = () => {
         <img src={logo} alt="logo" />
       </Link>
       <AuthNav />
-      <LoginForm />
     </StyledHeader>
   );
 };
