+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,5 +1,10 @@
 import React from 'react';
+import { LoginForm } from 'components/LoginForm/LoginForm';
 
 export default function LoginPage() {
-  return <div>LoginPage</div>;
+  return (
+    <div>
+      <LoginForm />
+    </div>
+  );
 }
